+++ ci/CHANGELOG.md
@@ -1,0 +1,2 @@
+# 0.1.0
+* Initial release
