+++ ci/README.md
@@ -1,0 +1,3 @@
+# ci
+
+A utility for the CI/CD integration of Fairwinds Insights.
