+++ ci/version.txt
@@ -1,0 +1,1 @@
+0.1.0
