+++ trivy/pkg/models/image.go
@@ -2,18 +2,22 @@ package models
 
 import "time"
 
+// Image represents a single container image to scan.
 type Image struct {
 	Name    string
 	ID      string
 	PullRef string
 	Owner   Resource
 }
+
+// Resource represents a Kubernetes resource
 type Resource struct {
 	Kind      string
 	Namespace string
 	Name      string
 }
 
+// ImageReport represents the results for a single resource.
 type ImageReport struct {
 	Name      string
 	ID        string
@@ -24,11 +28,13 @@ type ImageReport struct {
 	Report    []VulnerabilityList
 }
 
+// VulnerabilityList is the results from Trivy
 type VulnerabilityList struct {
 	Target          string
 	Vulnerabilities []Vulnerability
 }
 
+// Vulnerability is a single CVE or vulnerability.
 type Vulnerability struct {
 	Title            string
 	Description      string
@@ -40,11 +46,13 @@ type Vulnerability struct {
 	References       []string
 }
 
+// MinimizedReport is the results in a compressed format.
 type MinimizedReport struct {
 	Images          []ImageDetailsWithRefs
 	Vulnerabilities map[string]VulnerabilityDetails
 }
 
+// ImageDetailsWithRefs is the results of a scan for a resource with the vulnerabilities replaced with references.
 type ImageDetailsWithRefs struct {
 	ID        string
 	Name      string
@@ -55,11 +63,13 @@ type ImageDetailsWithRefs struct {
 	Report    []VulnerabilityRefList
 }
 
+// VulnerabilityRefList is a list of vulnerability references.
 type VulnerabilityRefList struct {
 	Target          string
 	Vulnerabilities []VulnerabilityInstance
 }
 
+// VulnerabilityDetails are the details of a vulnerability itself.
 type VulnerabilityDetails struct {
 	Title           string
 	Description     string
@@ -68,6 +78,7 @@ type VulnerabilityDetails struct {
 	References      []string
 }
 
+// VulnerabilityInstance is a single instance of a given vulnerability
 type VulnerabilityInstance struct {
 	InstalledVersion string
 	PkgName          string
